+++ src/main/java/net/silentchaos512/gems/SideProxy.java
@@ -154,6 +154,7 @@ class SideProxy {
         private void clientSetup(FMLClientSetupEvent event) {
             SilentGems.LOGGER.debug("Gems clientSetup");
             ModEntities.registerRenderers(event);
+            ModTileEntities.registerRenderers(event);
         }
 
         private static void registerContainers() {